requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
PyJWT==2.10.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
reportlab==4.0.9
PyPDF2==3.0.1

# Supabase e PostgreSQL
supabase==2.3.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Sistema de Assinatura Digital - Segurança Máxima
cryptography==42.0.5
pycryptodome==3.20.0
face-recognition==1.3.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==11.3.0
requests==2.31.0
hashlib-compat==1.0.1
base64-compat==1.0.1
